+++ ChangeLog.md
@@ -1,9 +1,11 @@
 # DirectoryBot Change Log
 #### DirectoryBot Version 2.1.0:
- - Some commands can now be used from private messages: `convert`, `credits`, `support`
+ - Some commands can now be used from private messages: `convert`, `credits`, `support`, `datapolicy`
  - The `credits` command now also returns the change log
  - `record` command messages are now deleted even if they encounter an error
  - New command: `raffle` picks a random user from a role, platform, and/or set of users
+ - Mentioning a role in a `newplatform` command sets that role as the platform role
+ - New command: `datapolicy` tells what kinds of data DirectoryBot collects and how it's used
 #### Example Version 1.0.0:
  - Change A
  - Change B